fix(): move exported functions in utils files when necessary (error observed by eslint-plugin-react-refresh)
