+++ src/components/menus/anchor-utils.ts
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) 2025, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+import { PopoverOrigin, PopoverPosition, PopoverProps, PopoverReference } from '@mui/material';
+
+export interface AnchorStatesType {
+    anchorEl?: PopoverProps['anchorEl'];
+    anchorReference?: PopoverReference;
+    anchorPosition?: PopoverPosition;
+    anchorOrigin?: PopoverOrigin;
+    transformOrigin?: PopoverOrigin;
+}
+
+export const defaultAnchorStates: AnchorStatesType = {
+    anchorReference: 'anchorPosition',
+};
