+++ src/components/menus/common-contextual-menu.tsx
@@ -8,19 +8,7 @@
 import { ReactNode } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { Edit as EditIcon } from '@mui/icons-material';
-import {
-    Divider,
-    ListItemIcon,
-    ListItemText,
-    Menu,
-    MenuItem,
-    MenuProps,
-    PopoverOrigin,
-    PopoverPosition,
-    PopoverProps,
-    PopoverReference,
-    styled,
-} from '@mui/material';
+import { Divider, ListItemIcon, ListItemText, Menu, MenuItem, MenuProps, styled } from '@mui/material';
 
 const StyledMenu = styled((props: MenuProps) => <Menu elevation={0} {...props} />)({
     '.MuiMenu-paper': {
@@ -43,18 +31,6 @@ export type MenuItemType =
 export interface CommonContextualMenuProps extends MenuProps {
     menuItems?: MenuItemType[];
 }
-
-export interface AnchorStatesType {
-    anchorEl?: PopoverProps['anchorEl'];
-    anchorReference?: PopoverReference;
-    anchorPosition?: PopoverPosition;
-    anchorOrigin?: PopoverOrigin;
-    transformOrigin?: PopoverOrigin;
-}
-
-export const defaultAnchorStates: AnchorStatesType = {
-    anchorReference: 'anchorPosition',
-};
 
 export default function CommonContextualMenu({ menuItems, ...menuProps }: Readonly<CommonContextualMenuProps>) {
     function makeMenuItem(
