+++ src/components/utils/directory-content-utils.ts
@@ -13,9 +13,10 @@ import { ColDef, IRowNode } from 'ag-grid-community';
 import type { ElementAttributes } from '@gridsuite/commons-ui';
 import { NameCellRenderer } from './renderers/name-cell-renderer';
 import { DescriptionCellRenderer } from './renderers/description-cell-renderer';
-import { getElementTypeTranslation, TypeCellRenderer } from './renderers/type-cell-renderer';
+import { TypeCellRenderer } from './renderers/type-cell-renderer';
 import { UserCellRenderer } from './renderers/user-cell-renderer';
 import { DateCellRenderer } from './renderers/date-cell-renderer';
+import { getElementTypeTranslation } from './translation-utils';
 
 export const formatMetadata = (
     data: ElementAttributes,
